+++ src/services/maakuntakunnat/actionTypes.js
@@ -1,0 +1,3 @@
+export const FETCH_MAAKUNTAKUNNAT_START = 'FETCH_MAAKUNTAKUNNAT_START'
+export const FETCH_MAAKUNTAKUNNAT_SUCCESS = 'FETCH_MAAKUNNAT_SUCCESS'
+export const FETCH_MAAKUNTAKUNNAT_FAILURE = 'FETCH_MAAKUNNAT_FAILURE'
